Bump the version of just we are using
